@@ -1,48 +1,68 @@
 plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
+    id("com.android.library")
 }
 
 android {
-    compileSdkVersion 31
+    compileSdk 33
 
     defaultConfig {
         minSdkVersion 24
-        targetSdkVersion 31
-        versionCode 1
-        versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
     }
 
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
+    lintOptions {
+        lintConfig file("lint.xml")
+    }
+
+    namespace "com.android.identity"
 }
 
 dependencies {
-    implementation("androidx.annotation:annotation:1.2.0")
-    implementation "androidx.biometric:biometric:$androidx_biometrics"
-    implementation "co.nstant.in:cbor:$cbor_version"
-    implementation "org.bouncycastle:bcprov-jdk15on:$bouncy_castle_version"
-    implementation("org.bouncycastle:bcpkix-jdk15on:1.56")
+    implementation("androidx.annotation:annotation:1.5.0")
+    implementation "androidx.biometric:biometric:1.2.0-alpha05"
+    implementation "co.nstant.in:cbor:0.9"
+    implementation "org.bouncycastle:bcprov-jdk15on:1.67"
+    implementation("org.bouncycastle:bcpkix-jdk15on:1.67")
+    implementation "com.android.volley:volley:1.2.1"
 
-    // JUnit 4 framework
-    testImplementation "junit:junit:$junit_version"
-    androidTestImplementation "androidx.test.ext:junit:$junit_test"
-    androidTestImplementation "androidx.test.espresso:espresso-core:$espresso_core_version"
+    testImplementation "androidx.test.espresso:espresso-core:3.4.0"
+    testImplementation "androidx.test.ext:junit:1.1.3"
+    testImplementation "junit:junit:4.13.2"
+    testImplementation "org.bouncycastle:bcprov-jdk15on:1.67"
 
-//    androidTestImplementation(libs.testExtJunit)
-//    androidTestImplementation(libs.testCore)
-//    androidTestImplementation(libs.testRunner)
-//    androidTestImplementation(libs.testRules)
-//    androidTestImplementation(libs.mockitoCore)
+    androidTestImplementation "androidx.test.ext:junit:1.1.3"
+    androidTestImplementation "androidx.test.espresso:espresso-core:3.4.0"
+}
+
+task generateApiDoc(type: Javadoc) {
+    source = android.sourceSets.main.java.srcDirs
+
+    ext.androidJar = "${android.sdkDirectory}/platforms/${android.compileSdkVersion}/android.jar"
+    classpath += project.files(android.getBootClasspath().join(File.pathSeparator))
+    classpath += files(ext.androidJar)
+
+    options.links("http://docs.oracle.com/javase/7/docs/api/");
+    options.links("http://d.android.com/reference/");
+
+    destinationDir = file("../docs/javadoc/")
+    failOnError false
+}
+
+tasks.withType(Test) {
+    testLogging {
+        events "standardOut", "passed", "skipped", "failed"
+    }
+}
+
+afterEvaluate {
+    generateApiDoc.classpath += files(android.libraryVariants.collect { variant ->
+        variant.javaCompileProvider.get().classpath.files
+    })
 }
