@@ -22,15 +22,18 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 
-@Database(entities = [Document::class], version = 1)
+@Database(
+    version = 2,
+    entities = [Document::class]
+)
 @TypeConverters(Converters::class)
 abstract class DocumentDatabase : RoomDatabase() {
-    abstract fun credentialDao(): DocumentDao
 
+    abstract fun credentialDao(): DocumentDao
 
     companion object {
 
-        private const val DATABASE_NAME = "document-db"
+        private const val DATABASE_NAME = "mdoc-document-database"
 
         // For Singleton instantiation
         @Volatile
@@ -42,10 +45,14 @@ abstract class DocumentDatabase : RoomDatabase() {
             }
         }
 
+        // TODO: According to https://developer.android.com/guide/topics/data/autobackup this
+        //   database ends up being backed up and restored. We should probably opt out of doing
+        //   that since all of these documents will reference HW-backed keys that are not backed
+        //   up and restored.
         private fun buildDatabase(context: Context): DocumentDatabase {
-            return Room.databaseBuilder(
-                context, DocumentDatabase::class.java, DATABASE_NAME
-            ).build()
+            return Room.databaseBuilder(context, DocumentDatabase::class.java, DATABASE_NAME)
+                .addMigrations(MigrationV1ToV2)
+                .build()
         }
     }
 }
