@@ -9,8 +9,10 @@ import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
+import com.android.identity.IdentityCredential
 import com.android.mdl.app.databinding.FragmentDocumentDetailBinding
 import com.android.mdl.app.document.Document
+import com.android.mdl.app.document.DocumentManager
 
 class DocumentDetailFragment : Fragment() {
     companion object {
@@ -24,11 +26,16 @@ class DocumentDetailFragment : Fragment() {
     // onDestroyView.
     private val binding get() = _binding!!
     private lateinit var document: Document
+    private var identityCredential: IdentityCredential? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         document = args.document
+
+        val documentManager = DocumentManager.getInstance(requireContext())
+        identityCredential = documentManager.getCredential(document)
+
         // Always return to main fragment
         requireActivity().onBackPressedDispatcher.addCallback(
             this,
@@ -46,10 +53,20 @@ class DocumentDetailFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentDocumentDetailBinding.inflate(inflater)
-        binding.fragment = this
-        binding.document = document
+        bindingUI()
 
         return binding.root
+    }
+
+    private fun bindingUI() {
+        binding.fragment = this
+        binding.document = document
+        binding.tvSelfSigned.text = if (document.selfSigned) "Yes" else "No"
+        binding.tvUserAuth.text = if (document.userAuthentication) "Yes" else "No"
+        binding.tvHardwareBacked.text = if (document.hardwareBacked) "Yes" else "No"
+        binding.tvMsoUsageCount.text =
+            identityCredential?.authenticationDataUsageCount.contentToString()
+        binding.btCheckUpdates.isEnabled = !document.selfSigned
     }
 
     fun onCheckUpdate() {
@@ -86,8 +103,24 @@ class DocumentDetailFragment : Fragment() {
         )
     }
 
+    fun onShowData() {
+        val direction = DocumentDetailFragmentDirections.navigateToDocumentData(document)
+        findNavController().navigate(direction)
+    }
+
     fun onDelete() {
-        if (document.serverUrl == null) {
+        if (document.selfSigned) {
+            val documentManager = DocumentManager.getInstance(requireContext())
+            documentManager.deleteCredentialByName(document.identityCredentialName)
+            Toast.makeText(
+                requireContext(), "Document deleted!",
+                Toast.LENGTH_SHORT
+            ).show()
+            findNavController().navigate(
+                DocumentDetailFragmentDirections.actionDocumentDetailFragmentToSelectDocumentFragment()
+            )
+            return
+        } else if (document.serverUrl == null) {
             // If server URL is null
             Toast.makeText(
                 requireContext(), "Document doesn't have server information",
@@ -101,6 +134,5 @@ class DocumentDetailFragment : Fragment() {
             )
         )
     }
-
 
 }
