@@ -6,23 +6,17 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
+import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.android.mdl.app.databinding.FragmentShareDocumentBinding
 import com.android.mdl.app.util.TransferStatus
 import com.android.mdl.app.viewmodel.ShareDocumentViewModel
 
-
 class ShareDocumentFragment : Fragment() {
-    companion object {
-        private const val LOG_TAG = "ShareDocumentFragment"
-    }
+
+    private val viewModel: ShareDocumentViewModel by viewModels()
 
     private var _binding: FragmentShareDocumentBinding? = null
-    private lateinit var vm: ShareDocumentViewModel
-
-    // This property is only valid between onCreateView and
-    // onDestroyView.
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -30,66 +24,67 @@ class ShareDocumentFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentShareDocumentBinding.inflate(inflater)
-        vm = ViewModelProvider(this).get(ShareDocumentViewModel::class.java)
-
-        binding.vm = vm
+        binding.vm = viewModel
         binding.fragment = this
-
-        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, callback)
-
+        requireActivity().onBackPressedDispatcher.addCallback(viewLifecycleOwner, onBackCallback)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        vm.startPresentation()
-
-        vm.getTransferStatus().observe(viewLifecycleOwner, {
+        viewModel.message.set("NFC tap with mdoc verifier device")
+        viewModel.getTransferStatus().observe(viewLifecycleOwner) {
             when (it) {
-                TransferStatus.ENGAGEMENT_READY -> {
-                    vm.message.set("Scan QR code or NFC tap with mdoc verifier device")
-                    vm.setDeviceEngagement()
+                TransferStatus.QR_ENGAGEMENT_READY -> {
+                    viewModel.message.set("Scan QR code with mdoc verifier device")
+                    viewModel.showQrCode()
                 }
+
                 TransferStatus.CONNECTED -> {
-                    vm.message.set("Connected!")
-                    findNavController().navigate(
-                        ShareDocumentFragmentDirections.actionShareDocumentFragmentToTransferDocumentFragment()
-                    )
+                    viewModel.message.set("Connected!")
+                    val destination = ShareDocumentFragmentDirections.toTransferDocumentFragment()
+                    findNavController().navigate(destination)
                 }
+
                 TransferStatus.REQUEST -> {
-                    vm.message.set("Request received!")
+                    viewModel.message.set("Request received!")
                 }
+
                 TransferStatus.DISCONNECTED -> {
-                    vm.message.set("Disconnected!")
-                    findNavController().navigate(
-                        ShareDocumentFragmentDirections.actionShareDocumentFragmentToSelectDocumentFragment()
-                    )
+                    viewModel.message.set("Disconnected!")
+                    findNavController().navigateUp()
                 }
+
                 TransferStatus.ERROR -> {
-                    vm.message.set("Error on presentation!")
+                    viewModel.message.set("Error on presentation!")
                 }
+
                 TransferStatus.ENGAGEMENT_DETECTED -> {
-                    vm.message.set("Engagement detected!")
+                    viewModel.message.set("Engagement detected!")
                 }
+
                 TransferStatus.CONNECTING -> {
-                    vm.message.set("Connecting...")
+                    viewModel.message.set("Connecting...")
                 }
+
+                else -> {}
             }
         }
-        )
     }
 
-    // This callback will only be called when MyFragment is at least Started.
-    var callback = object : OnBackPressedCallback(true /* enabled by default */) {
+    override fun onResume() {
+        super.onResume()
+        viewModel.triggerQrEngagement()
+    }
+
+    private val onBackCallback = object : OnBackPressedCallback(true) {
         override fun handleOnBackPressed() {
             onCancel()
         }
     }
 
     fun onCancel() {
-        vm.cancelPresentation()
-        findNavController().navigate(
-            ShareDocumentFragmentDirections.actionShareDocumentFragmentToSelectDocumentFragment()
-        )
+        viewModel.cancelPresentation()
+        findNavController().navigateUp()
     }
 }
