@@ -8,7 +8,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
-import androidx.security.identity.PersonalizationData
+import com.android.identity.PersonalizationData
 import com.android.mdl.app.databinding.FragmentUpdateCheckBinding
 import com.android.mdl.app.document.Document
 import com.android.mdl.app.document.DocumentManager
