@@ -4,9 +4,9 @@ import android.content.Context;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
-import androidx.security.identity.AccessControlProfile;
-import androidx.security.identity.AccessControlProfileId;
-import androidx.security.identity.PersonalizationData;
+import com.android.identity.AccessControlProfile;
+import com.android.identity.AccessControlProfileId;
+import com.android.identity.PersonalizationData;
 
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
