@@ -1,11 +1,12 @@
 package com.android.mdl.app.util
 
 enum class TransferStatus {
-    ENGAGEMENT_READY,
+    QR_ENGAGEMENT_READY,
     ENGAGEMENT_DETECTED,
     CONNECTING,
     CONNECTED,
     REQUEST,
+    REQUEST_SERVED,
     DISCONNECTED,
     ERROR
 }
