@@ -3,21 +3,21 @@ package com.android.mdl.app.viewmodel
 import android.app.Application
 import android.util.Log
 import android.view.View
-import androidx.biometric.BiometricPrompt.CryptoObject
 import androidx.databinding.ObservableField
 import androidx.databinding.ObservableInt
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
-import androidx.security.identity.Constants.DEVICE_RESPONSE_STATUS_OK
-import androidx.security.identity.DeviceRequestParser
-import androidx.security.identity.DeviceResponseGenerator
-import androidx.security.identity.InvalidRequestMessageException
+import androidx.lifecycle.MutableLiveData
+import com.android.identity.Constants.DEVICE_RESPONSE_STATUS_OK
+import com.android.identity.DeviceRequestParser
+import com.android.identity.DeviceResponseGenerator
 import com.android.mdl.app.R
+import com.android.mdl.app.authconfirmation.RequestedDocumentData
+import com.android.mdl.app.authconfirmation.SignedPropertiesCollection
 import com.android.mdl.app.document.Document
 import com.android.mdl.app.document.DocumentManager
 import com.android.mdl.app.transfer.TransferManager
 import com.android.mdl.app.util.TransferStatus
-import java.util.NoSuchElementException
 
 class TransferDocumentViewModel(val app: Application) : AndroidViewModel(app) {
 
@@ -25,89 +25,125 @@ class TransferDocumentViewModel(val app: Application) : AndroidViewModel(app) {
         private const val LOG_TAG = "TransferDocumentViewModel"
     }
 
-    var inProgress = ObservableInt(View.GONE)
-    var documentsSent = ObservableField<String>()
-    private var documentsCount = 0
-
     private val transferManager = TransferManager.getInstance(app.applicationContext)
     private val documentManager = DocumentManager.getInstance(app.applicationContext)
+    private val signedProperties = SignedPropertiesCollection()
+    private val requestedProperties = mutableListOf<RequestedDocumentData>()
+    private val closeConnectionMutableLiveData = MutableLiveData<Boolean>()
+    private val selectedDocuments = mutableListOf<Document>()
 
-    fun getTransferStatus(): LiveData<TransferStatus> = transferManager.getTransferStatus()
+    var inProgress = ObservableInt(View.GONE)
+    var documentsSent = ObservableField<String>()
+    val connectionClosedLiveData: LiveData<Boolean> = closeConnectionMutableLiveData
+
+    fun getTransferStatus(): LiveData<TransferStatus> =
+        transferManager.getTransferStatus()
 
     fun getRequestedDocuments(): Collection<DeviceRequestParser.DocumentRequest> =
-        transferManager.getDeviceRequest().docRequests
+        transferManager.documentRequests()
 
     fun getDocuments() = documentManager.getDocuments()
 
-    fun getEntryNames(): Map<Document, List<String>> {
-        val documents = mutableMapOf<Document, List<String>>()
-        val docRequests =
-            transferManager.getDeviceRequest().docRequests
-
-        docRequests.forEach { doc ->
-            try {
-                val entryNames = mutableListOf<String>()
-                doc.namespaces.forEach { ns ->
-                    entryNames.addAll(doc.getEntryNames(ns))
-                }
-                documents[documentManager.getDocuments().first { it.docType == doc.docType }] =
-                    entryNames
-            } catch (e : NoSuchElementException) {
-                Log.w(LOG_TAG, "No document for docType " + doc.docType)
-            }
-        }
-        return documents
-    }
+    fun getSelectedDocuments() = selectedDocuments
 
     fun getCryptoObject() = transferManager.getCryptoObject()
 
-    @Throws(InvalidRequestMessageException::class)
-    fun sendResponse(): CryptoObject? {
-        inProgress.set(View.VISIBLE)
-        // Currently we don't care about the request, for now we just send the mdoc we have
-        // without any regard to what the reader actually requested...
-        //
-        val documents = getDocuments()
+    fun requestedProperties() = requestedProperties
+
+    fun closeConnection() {
+        cleanUp()
+        closeConnectionMutableLiveData.value = true
+    }
+
+    fun addDocumentForSigning(document: RequestedDocumentData) {
+        signedProperties.addNamespace(document)
+    }
+
+    fun toggleSignedProperty(namespace: String, property: String) {
+        signedProperties.toggleProperty(namespace, property)
+    }
+
+    fun createSelectedItemsList() {
+        val ownDocuments = getSelectedDocuments()
         val requestedDocuments = getRequestedDocuments()
-        val response = DeviceResponseGenerator(DEVICE_RESPONSE_STATUS_OK)
-        requestedDocuments.forEach { reqDoc ->
+        val result = mutableListOf<RequestedDocumentData>()
+        requestedDocuments.forEach { requestedDocument ->
             try {
-                val doc = documents.first { it.docType == reqDoc.docType }
-                val issuerSignedEntriesToRequest = mutableMapOf<String, Collection<String>>()
-                reqDoc.namespaces.forEach { ns ->
-                    issuerSignedEntriesToRequest[ns] = reqDoc.getEntryNames(ns)
-                }
-                val authNeeded = transferManager.addDocumentToResponse(
-                    doc.identityCredentialName,
-                    doc.docType,
-                    issuerSignedEntriesToRequest,
-                    response,
-                    reqDoc.readerAuth,
-                    reqDoc.itemsRequest
-                )
-                if (authNeeded) {
-                    inProgress.set(View.GONE)
-                    inProgress.notifyChange()
-                    return transferManager.getCryptoObject()
-                }
-            } catch (e : NoSuchElementException) {
-                Log.w(LOG_TAG, "No document for docType " + reqDoc.docType)
+                val ownDocument = ownDocuments.first { it.docType == requestedDocument.docType }
+                val issuerSignedEntriesToRequest = requestedPropertiesFrom(requestedDocument)
+                result.addAll(issuerSignedEntriesToRequest.map {
+                    RequestedDocumentData(
+                        ownDocument.userVisibleName,
+                        it.key,
+                        ownDocument.identityCredentialName,
+                        false,
+                        it.value,
+                        requestedDocument
+                    )
+                })
+            } catch (e: NoSuchElementException) {
+                Log.w(LOG_TAG, "No document for docType " + requestedDocument.docType)
             }
         }
+        requestedProperties.addAll(result)
+    }
 
+    fun sendResponseForSelection(): Boolean {
+        val propertiesToSend = signedProperties.collect()
+        val response = DeviceResponseGenerator(DEVICE_RESPONSE_STATUS_OK)
+        propertiesToSend.forEach { signedDocument ->
+            try {
+                val issuerSignedEntries = with(signedDocument) {
+                    mutableMapOf(namespace to signedProperties)
+                }
+                val authNeeded = transferManager.addDocumentToResponse(
+                    signedDocument.identityCredentialName,
+                    signedDocument.documentType,
+                    issuerSignedEntries,
+                    response,
+                    signedDocument.readerAuth,
+                    signedDocument.itemsRequest
+                )
+                if (authNeeded) {
+                    return false
+                }
+            } catch (e: NoSuchElementException) {
+                Log.w(LOG_TAG, "No requestedDocument for " + signedDocument.documentType)
+            }
+        }
         transferManager.sendResponse(response.generate())
-        documentsCount++
+        transferManager.setResponseServed()
+        val documentsCount = propertiesToSend.count()
         documentsSent.set(app.getString(R.string.txt_documents_sent, documentsCount))
-        inProgress.set(View.GONE)
-        return null
+        cleanUp()
+        return true
     }
 
     fun cancelPresentation(
         sendSessionTerminationMessage: Boolean,
         useTransportSpecificSessionTermination: Boolean
-    ) =
+    ) {
         transferManager.stopPresentation(
             sendSessionTerminationMessage,
             useTransportSpecificSessionTermination
         )
+    }
+
+    private fun requestedPropertiesFrom(
+        requestedDocument: DeviceRequestParser.DocumentRequest
+    ): MutableMap<String, Collection<String>> {
+        val result = mutableMapOf<String, Collection<String>>()
+        requestedDocument.namespaces.forEach { namespace ->
+            val list = result.getOrDefault(namespace, ArrayList())
+            (list as ArrayList).addAll(requestedDocument.getEntryNames(namespace))
+            result[namespace] = list
+        }
+        return result
+    }
+
+    private fun cleanUp() {
+        requestedProperties.clear()
+        signedProperties.clear()
+        selectedDocuments.clear()
+    }
 }
