@@ -23,8 +23,6 @@ import com.android.mdl.appreader.transfer.TransferManager
 import com.android.mdl.appreader.util.TransferStatus
 import com.budiyev.android.codescanner.CodeScanner
 import com.budiyev.android.codescanner.DecodeCallback
-import java.util.*
-
 
 /**
  * A simple [Fragment] subclass as the second destination in the navigation.
@@ -88,46 +86,51 @@ class DeviceEngagementFragment : Fragment() {
 
         binding.csScanner.setOnClickListener { mCodeScanner?.startPreview() }
 
-        transferManager.getTransferStatus().observe(viewLifecycleOwner, {
+        transferManager.getTransferStatus().observe(viewLifecycleOwner) {
             when (it) {
                 TransferStatus.ENGAGED -> {
                     Log.d(LOG_TAG, "Device engagement received")
                     onDeviceEngagementReceived()
                 }
+
                 TransferStatus.CONNECTED -> {
                     Log.d(LOG_TAG, "Device connected")
                     Toast.makeText(
                         requireContext(), "Error invalid callback connected",
                         Toast.LENGTH_SHORT
                     ).show()
-                    findNavController().navigate(R.id.action_Transfer_to_RequestOptions)
+                    findNavController().navigate(R.id.action_ScanDeviceEngagement_to_RequestOptions)
                 }
+
                 TransferStatus.RESPONSE -> {
                     Log.d(LOG_TAG, "Device response received")
                     Toast.makeText(
                         requireContext(), "Error invalid callback response",
                         Toast.LENGTH_SHORT
                     ).show()
-                    findNavController().navigate(R.id.action_Transfer_to_RequestOptions)
+                    findNavController().navigate(R.id.action_ScanDeviceEngagement_to_RequestOptions)
                 }
+
                 TransferStatus.DISCONNECTED -> {
                     Log.d(LOG_TAG, "Device disconnected")
                     Toast.makeText(
                         requireContext(), "Device disconnected",
                         Toast.LENGTH_SHORT
                     ).show()
-                    findNavController().navigate(R.id.action_Transfer_to_RequestOptions)
+                    findNavController().navigate(R.id.action_ScanDeviceEngagement_to_RequestOptions)
                 }
+
                 TransferStatus.ERROR -> {
                     Log.d(LOG_TAG, "Error received")
                     Toast.makeText(
                         requireContext(), "Error connecting to holder",
                         Toast.LENGTH_SHORT
                     ).show()
-                    findNavController().navigate(R.id.action_Transfer_to_RequestOptions)
+                    findNavController().navigate(R.id.action_ScanDeviceEngagement_to_RequestOptions)
                 }
+                else -> {}
             }
-        })
+        }
 
         binding.btCancel.setOnClickListener {
             findNavController().navigate(R.id.action_ScanDeviceEngagement_to_RequestOptions)
@@ -206,7 +209,7 @@ class DeviceEngagementFragment : Fragment() {
     }
 
     private fun onDeviceEngagementReceived() {
-        if (transferManager.availableMdocAddresses?.size == 1) {
+        if (transferManager.availableMdocConnectionMethods?.size == 1) {
             findNavController().navigate(
                 DeviceEngagementFragmentDirections.actionScanDeviceEngagementToTransfer(
                     args.requestDocumentList
