@@ -2,9 +2,11 @@ package com.android.mdl.appreader.util
 
 import android.content.Context
 import androidx.preference.PreferenceManager
-import androidx.security.identity.Constants
+import com.android.identity.Constants
 
 object PreferencesHelper {
+    private const val BLE_DATA_L2CAP = "ble_l2cap"
+    private const val BLE_CLEAR_CACHE = "ble_clear_cache"
     private const val READER_AUTHENTICATION = "reader_authentication"
     private const val LOG_INFO = "log_info"
     private const val LOG_DEVICE_ENGAGEMENT = "log_device_engagement"
@@ -12,32 +14,16 @@ object PreferencesHelper {
     private const val LOG_TRANSPORT = "log_transport"
     private const val LOG_TRANSPORT_VERBOSE = "log_transport_verbose"
 
-    fun getLoggingFlags(context: Context): Int {
-        var flags = 0
-        if (PreferenceManager.getDefaultSharedPreferences(context).getBoolean(LOG_INFO, true)) {
-            flags += Constants.LOGGING_FLAG_INFO
-        }
-        if (PreferenceManager.getDefaultSharedPreferences(context)
-                .getBoolean(LOG_DEVICE_ENGAGEMENT, true)
-        ) {
-            flags += Constants.LOGGING_FLAG_DEVICE_ENGAGEMENT
-        }
-        if (PreferenceManager.getDefaultSharedPreferences(context)
-                .getBoolean(LOG_SESSION_MESSAGES, true)
-        ) {
-            flags += Constants.LOGGING_FLAG_SESSION_MESSAGES
-        }
-        if (PreferenceManager.getDefaultSharedPreferences(context)
-                .getBoolean(LOG_TRANSPORT, true)
-        ) {
-            flags += Constants.LOGGING_FLAG_TRANSPORT_SPECIFIC
-        }
-        if (PreferenceManager.getDefaultSharedPreferences(context)
-                .getBoolean(LOG_TRANSPORT_VERBOSE, true)
-        ) {
-            flags += Constants.LOGGING_FLAG_TRANSPORT_SPECIFIC_VERBOSE
-        }
-        return flags
+    fun isBleL2capEnabled(context: Context): Boolean {
+        return PreferenceManager.getDefaultSharedPreferences(context).getBoolean(
+            BLE_DATA_L2CAP, false
+        )
+    }
+
+    fun isBleClearCacheEnabled(context: Context): Boolean {
+        return PreferenceManager.getDefaultSharedPreferences(context).getBoolean(
+            BLE_CLEAR_CACHE, false
+        )
     }
 
     fun getReaderAuth(context: Context): String {
